mcp-outlook-client/requirements.txt
# MCP Client Dependencies
mcp>=0.9.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Email sending
secure-smtplib>=0.1.1

# Optional: For better async support
asyncio>=3.4.3